implemented threads and futex wait. ARM timer is next so nanotime and clock_gettime() will work.
There is a global pointer to the current trapframe and threads can
call RecordTrapframe() to snapshot their execution state. To switch
threads, point the current trapframe somewhere else and call
ReplayTrapframe(). Clone records a trapframe in an empty thread struct
and sets its state to runnable so the scheduler can pick it up.